% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earliest_dob.R
\name{earliest_dob}
\alias{earliest_dob}
\title{Calculate earliest possible date of birth given an age and date}
\usage{
earliest_dob(age, date)
}
\arguments{
\item{age}{Numeric vector of ages}

\item{date}{Date vector of observation dates}
}
\value{
Date vector of earliest possible dates of birth
}
\description{
Calculate earliest possible date of birth given an age and date
}
\details{
The earliest possible date of birth is if this is the person's last day
as an \code{age} year old. Therefore, the earliest possible date is tomorrow's
date minus \code{age + 1} years.
}
