% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dob.R
\name{estimate_dob}
\alias{estimate_dob}
\title{Estimate date of birth from ages at known dates}
\usage{
estimate_dob(age, date)
}
\arguments{
\item{age}{Numeric vector of ages}

\item{date}{Date vector of observation dates}
}
\value{
Interval from earliest to latest possible date of birth
}
\description{
Estimate date of birth from ages at known dates
}
