% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_dob.R
\name{latest_dob}
\alias{latest_dob}
\title{Calculate latest possible date of birth given an age and date}
\usage{
latest_dob(age, date)
}
\arguments{
\item{age}{Numeric vector of ages}

\item{date}{Date vector of observation dates}
}
\value{
Date vector of latest possible dates of birth
}
\description{
Calculate latest possible date of birth given an age and date
}
\details{
The latest possible date of birth is if the person turned this age
today.
}
